Origination
This file is mostly notes for code that allows you to deploy a transaction between two accounts on ganache, a private blockchain. We can manually create, sign, send, and view the transaction record using web3. The results are viewed on ganache. This could be executed on mainnet as well but would require expenditure and transfer of real ether.